+++ src/js/videojs.record.js
@@ -20,10 +20,11 @@ import defaultKeyHandler from './hot-keys';
 import pluginDefaultOptions from './defaults';
 import formatTime from './utils/format-time';
 import setSrcObject from './utils/browser-shim';
+import compareVersion from './utils/compare-version';
 import {detectBrowser} from './utils/detect-browser';
 
 import {getAudioEngine, isAudioPluginActive, getVideoEngine, getConvertEngine, getAnimationEngine} from './engine/engine-loader';
-import {IMAGE_ONLY, AUDIO_ONLY, VIDEO_ONLY, AUDIO_VIDEO, ANIMATION, SCREEN_ONLY, getRecorderMode} from './engine/record-mode';
+import {IMAGE_ONLY, AUDIO_ONLY, VIDEO_ONLY, AUDIO_VIDEO, AUDIO_SCREEN, ANIMATION, SCREEN_ONLY, getRecorderMode} from './engine/record-mode';
 
 const Plugin = videojs.getPlugin('plugin');
 const Player = videojs.getComponent('Player');
@@ -68,6 +69,7 @@ class Record extends Plugin {
 
         // add device button with icon based on type
         let deviceIcon = 'av-perm';
+
         switch (this.getRecordType()) {
             case IMAGE_ONLY:
             case VIDEO_ONLY:
@@ -79,6 +81,9 @@ class Record extends Plugin {
                 break;
             case SCREEN_ONLY:
                 deviceIcon = 'screen-perm';
+                break;
+            case AUDIO_SCREEN:
+                deviceIcon = 'sv-perm';
                 break;
         }
 
@@ -113,13 +118,15 @@ class Record extends Plugin {
         player.recordToggle = new RecordToggle(player, options);
         player.recordToggle.hide();
 
-        // add picture-in-picture toggle button
-        player.pipToggle = new PictureInPictureToggle(player, options);
-        player.pipToggle.hide();
+        // add picture-in-picture toggle button for older video.js versions
+        if (videojs.VERSION === undefined || compareVersion(videojs.VERSION, '7.6.0') === -1) {
+            player.pipToggle = new PictureInPictureToggle(player, options);
+            player.pipToggle.hide();
+        }
 
         // picture-in-picture
         if (this.pictureInPicture === true) {
-            // dfine Picture-in-Picture event handlers once
+            // define Picture-in-Picture event handlers once
             this.onEnterPiPHandler = this.onEnterPiP.bind(this);
             this.onLeavePiPHandler = this.onLeavePiP.bind(this);
         }
@@ -127,7 +134,10 @@ class Record extends Plugin {
         // exclude custom UI elements
         if (this.player.options_.controlBar) {
             let customUIElements = ['deviceButton', 'recordIndicator',
-                'cameraButton', 'recordToggle', 'pipToggle'];
+                'cameraButton', 'recordToggle'];
+            if (player.pipToggle) {
+                customUIElements.push('pipToggle');
+            }
             customUIElements.forEach((element) => {
                 if (this.player.options_.controlBar[element] !== undefined) {
                     this.player[element].layoutExclude = true;
@@ -209,7 +219,14 @@ class Record extends Plugin {
         this.player.controlBar.el().insertBefore(
             this.player.recordToggle.el(),
             this.player.controlBar.el().firstChild);
-        this.player.controlBar.addChild(this.player.pipToggle);
+        if (this.player.controlBar.pictureInPictureToggle === undefined) {
+            // add custom PiP toggle
+            this.player.controlBar.addChild(this.player.pipToggle);
+        } else {
+            // use video.js PiP toggle
+            this.player.pipToggle = this.player.controlBar.pictureInPictureToggle;
+            this.player.pipToggle.hide();
+        }
 
         // get rid of unused controls
         if (this.player.controlBar.remainingTimeDisplay !== undefined) {
@@ -234,6 +251,7 @@ class Record extends Plugin {
             case AUDIO_VIDEO:
             case ANIMATION:
             case SCREEN_ONLY:
+            case AUDIO_SCREEN:
                 // customize controls
                 this.player.bigPlayButton.hide();
 
@@ -361,7 +379,7 @@ class Record extends Plugin {
         // check for support because some browsers still do not support
         // getDisplayMedia or getUserMedia (like Chrome iOS, see:
         // https://bugs.chromium.org/p/chromium/issues/detail?id=752458)
-        if (this.getRecordType() === SCREEN_ONLY) {
+        if (this.getRecordType() === SCREEN_ONLY || this.getRecordType() === AUDIO_SCREEN) {
             if (navigator.mediaDevices === undefined ||
                 navigator.mediaDevices.getDisplayMedia === undefined) {
                 this.player.trigger(Event.ERROR,
@@ -445,6 +463,25 @@ class Record extends Plugin {
                 }).then(
                     this.onDeviceReady.bind(this)
                 ).catch(
+                    this.onDeviceError.bind(this)
+                );
+                break;
+
+            case AUDIO_SCREEN:
+                // setup camera and microphone
+                this.mediaType = {
+                    audio: (this.audioRecorderType === AUTO) ? true : this.audioRecorderType,
+                    video: (this.videoRecorderType === AUTO) ? true : this.videoRecorderType
+                };
+                navigator.mediaDevices.getDisplayMedia({
+                    video: true // This needs to be true for Firefox to work
+                }).then(screenStream => {
+                    navigator.mediaDevices.getUserMedia({ audio:this.recordAudio }).then((mic) => {
+                        // Join microphone track with screencast stream (order matters)
+                        screenStream.addTrack(mic.getTracks()[0]);
+                        this.onDeviceReady.bind(this)(screenStream);
+                    });
+                }).catch(
                     this.onDeviceError.bind(this)
                 );
                 break;
@@ -760,6 +797,7 @@ class Record extends Plugin {
 
                 case VIDEO_ONLY:
                 case AUDIO_VIDEO:
+                case AUDIO_SCREEN:
                 case SCREEN_ONLY:
                     // preview video stream in video element
                     this.startVideoPreview();
@@ -798,6 +836,7 @@ class Record extends Plugin {
 
                 case VIDEO_ONLY:
                 case AUDIO_VIDEO:
+                case AUDIO_SCREEN:
                 case ANIMATION:
                 case SCREEN_ONLY:
                     // wait for media stream on video element to actually load
@@ -987,6 +1026,7 @@ class Record extends Plugin {
 
             case VIDEO_ONLY:
             case AUDIO_VIDEO:
+            case AUDIO_SCREEN:
             case SCREEN_ONLY:
                 // pausing the player so we can visualize the recorded data
                 // will trigger an async video.js 'pause' event that we
@@ -1008,7 +1048,7 @@ class Record extends Plugin {
                         this.playbackTimeUpdate);
 
                     // unmute local audio during playback
-                    if (this.getRecordType() === AUDIO_VIDEO) {
+                    if (this.getRecordType() === AUDIO_VIDEO || this.getRecordType() === AUDIO_SCREEN) {
                         this.mediaElement.muted = false;
 
                         // show the volume bar when it's unmuted
@@ -1120,6 +1160,7 @@ class Record extends Plugin {
 
             case VIDEO_ONLY:
             case AUDIO_VIDEO:
+            case AUDIO_SCREEN:
             case ANIMATION:
             case SCREEN_ONLY:
                 if (this.player.controlBar.currentTimeDisplay &&
@@ -1164,6 +1205,7 @@ class Record extends Plugin {
 
             case VIDEO_ONLY:
             case AUDIO_VIDEO:
+            case AUDIO_SCREEN:
             case ANIMATION:
             case SCREEN_ONLY:
                 // update duration display component
@@ -1193,6 +1235,7 @@ class Record extends Plugin {
             case IMAGE_ONLY:
             case VIDEO_ONLY:
             case AUDIO_VIDEO:
+            case AUDIO_SCREEN:
             case ANIMATION:
             case SCREEN_ONLY:
                 if (url instanceof Blob || url instanceof File) {
@@ -1372,6 +1415,7 @@ class Record extends Plugin {
      */
     muteTracks(mute) {
         if ((this.getRecordType() === AUDIO_ONLY ||
+            this.getRecordType() === AUDIO_SCREEN ||
             this.getRecordType() === AUDIO_VIDEO) &&
             this.stream.getAudioTracks().length > 0) {
             this.stream.getAudioTracks()[0].enabled = !mute;