+++ CHANGES.md
@@ -1,16 +1,19 @@
 videojs-record changelog
 ========================
 
-3.8.0 - unreleased
+3.8.0 - 2019/07/16
 ------------------
 
 - New ffmpeg.js plugin: convert recorded data into other audio/video file formats
   in the browser (#201)
+- Support for capturing screen + audio (#385 by @tony)
 - Support for specifying third-party plugin settings using the `pluginLibraryOptions`
   option (#383)
 - New options: `videoBitRate` and `videoFrameRate` (currently only used in the
   webm-wasm plugin)
 - New event: `startConvert` (used in ts-ebml and ffmpeg.js plugins) (#201)
+- Support for video.js Picture-In-Picture API (available in video.js >= 7.6.0). Tries
+  to use `PictureInPictureToggle`, otherwise fallback to own implementation (#381)
 - Fix `RecordRTC.MediaStreamRecorder` import when using the `timeSlice` option
 - Examples: add `timeSlice` example demonstrating use of `timestamp` event
 - Bump required version for videojs-wavesurfer to 2.9.0 or newer for wavesurfer.js